Refactor canvas behavior into focused components
